advice-lat/requirements.txt
fastapi==0.61.0
uvicorn==0.11.8
grpcio==1.33.2
protobuf==3.14.0
grpcio-tools==1.33.2
jinja2==2.11.2
aiofiles==0.6.0
opencv-python==4.5.3.56
Pillow